+++ pylops/basicoperators/laplacian.py
@@ -74,7 +74,7 @@ class Laplacian(LinearOperator):
         self.edge = edge
         self.kind = kind
         super().__init__(Op=self._calc_l2op(dims=dims, axes=axes, sampling=sampling, edge=edge, kind=kind, dtype=dtype,
-                                            weights=weights), dtype=dtype, name=name)
+                                            weights=weights), name=name)
 
     def _matvec(self, x: NDArray) -> NDArray:
         return super()._matvec(x)
